move arrow_io_source to separate header
